refactor/include image quality to reset target
